mcp-bridge/requirements.txt
flask==3.0.0
requests==2.31.0
sseclient-py==1.8.0